+++ drf_spectacular_jsonapi/schemas/plumbing.py
@@ -5,6 +5,8 @@ from rest_framework_json_api.utils import (format_field_name,
                                            get_related_resource_type,
                                            get_resource_type_from_serializer)
 
+from drf_spectacular_jsonapi.schemas.utils import get_primary_key_of_serializer
+
 
 def build_json_api_relationship_object(field):
     schema = {
@@ -53,9 +55,9 @@ def build_json_api_resource_object(schema, serializer, method):
     required = []
     attributes = {}
     relationships = {}
-    result = {
+    resource_object_schema = {
         "type": "object",
-        "required": ["type", "id"],
+        "required": ["type"],
         "additionalProperties": False,
         "properties": {
             "type": {
@@ -63,7 +65,6 @@ def build_json_api_resource_object(schema, serializer, method):
                 "description": _("The [type](https://jsonapi.org/format/#document-resource-object-identification) member is used to describe resource objects that share common attributes and relationships."),
                 "enum": [get_resource_type_from_serializer(serializer=serializer)]
             },
-            "id": schema["properties"]["id"],
             # TODO:
             # "links": {
             #     "type": "object",
@@ -71,18 +72,32 @@ def build_json_api_resource_object(schema, serializer, method):
             # },
         },
     }
+    pk_name = get_primary_key_of_serializer(serializer=serializer)
 
-    if method == "POST" and serializer.fields["id"].read_only:
-        # no id is needed on creating resources. Exception: https://jsonapi.org/format/#crud-creating-client-ids
-        result["required"].remove("id")
-        del result["properties"]["id"]
-    elif method == "POST" and not serializer.fields["id"].read_only or method == "PATCH":
+    if method == "PATCH" or method == "GET" or pk_name and method == "POST" and not serializer.fields[pk_name].read_only:
+        # case 1: PATCH:
+        # The PATCH request MUST include a single resource object as primary data.
+        # The resource object MUST contain type and id members.
+
+        # case 2: "GET"
+        # If method == "GET" this resource object schema shall be build for an response body schema definition.
         # id is required
-        del result["properties"]["id"]["readOnly"]
+
+        # case 3: "POST" with client id see: https://jsonapi.org/format/#crud-creating-client-ids
+        resource_object_schema["required"].append("id")
+
+        # {} is a shorthand syntax for an arbitrary-type: see https://swagger.io/docs/specification/data-models/data-types/#any
+        resource_object_schema["properties"]["id"] = schema["properties"][pk_name] if pk_name else {
+        }
+
+        is_read_only = resource_object_schema["properties"]["id"].get(
+            "readOnly", None)
+        if is_read_only:
+            del resource_object_schema["properties"]["id"]["readOnly"]
 
     for field in serializer.fields.values():
         field_schema = schema["properties"][field.field_name]
-        if field.field_name == "id":
+        if field.field_name == pk_name:
             # id field shall not be part of the attributes
             continue
         if isinstance(field, serializers.HyperlinkedIdentityField):
@@ -111,17 +126,17 @@ def build_json_api_resource_object(schema, serializer, method):
         attributes[format_field_name(field.field_name)] = field_schema
 
     if attributes:
-        result["properties"]["attributes"] = {
+        resource_object_schema["properties"]["attributes"] = {
             "type": "object",
             "properties": attributes,
         }
         if required:
-            result["properties"]["attributes"]["required"] = required
+            resource_object_schema["properties"]["attributes"]["required"] = required
 
     if relationships:
-        result["properties"]["relationships"] = {
+        resource_object_schema["properties"]["relationships"] = {
             "type": "object",
             "properties": relationships,
         }
 
-    return result
+    return resource_object_schema