+++ tests/serializers.py
@@ -1,12 +1,15 @@
 from django.utils.translation import gettext_lazy as _
+from rest_framework.fields import CharField
 from rest_framework_json_api.relations import ResourceRelatedField
-from rest_framework_json_api.serializers import ModelSerializer
+from rest_framework_json_api.serializers import ModelSerializer, Serializer
 
-from .models import Album, Song
+from .models import Album, Song, User
 
 __all__ = [
     "SongSerializer",
-    "AlbumSerializer"
+    "AlbumSerializer",
+    "SessionCreateSerializer",
+    "UserSerializer"
 ]
 
 
@@ -37,4 +40,33 @@ class AlbumSerializer(ModelSerializer):
 
     class Meta:
         model = Album
+        fields = "__all__"
+
+
+class PasswordField(CharField):
+    def __init__(self, *args, **kwargs):
+        kwargs.setdefault('style', {})
+
+        kwargs['style']['input_type'] = 'password'
+        kwargs['write_only'] = True
+
+        super().__init__(*args, **kwargs)
+
+
+class SessionCreateSerializer(Serializer):
+    """Simple non model serializer"""
+    username = CharField(
+        label=_("username"),
+    )
+    password = PasswordField(
+        label=_("password"),
+    )
+
+    class Meta:
+        resource_name = 'SessionCreate'
+
+
+class UserSerializer(ModelSerializer):
+    class Meta:
+        model = User
         fields = "__all__"