+++ tests/views.py
@@ -1,7 +1,11 @@
+from django.contrib.auth import login
+from rest_framework.generics import CreateAPIView
+from rest_framework.response import Response
 from rest_framework_json_api.views import ModelViewSet
 
-from .models import Album, Song
-from .serializers import AlbumSerializer, SongSerializer
+from .models import Album, Song, User
+from .serializers import (AlbumSerializer, SessionCreateSerializer,
+                          SongSerializer, UserSerializer)
 
 
 class AlbumModelViewset(ModelViewSet):
@@ -20,3 +24,8 @@ class SongModelViewset(ModelViewSet):
     """ """
     serializer_class = SongSerializer
     queryset = Song.objects.none()
+
+
+class UserModelViewset(ModelViewSet):
+    serializer_class = UserSerializer
+    queryset = User.objects.none()